+++ src/Geomancy/IVec4.hs
@@ -21,6 +21,7 @@ import Data.Int (Int32)
 import Data.MonoTraversable (Element, MonoFunctor(..), MonoPointed(..))
 import Foreign (Storable(..))
 import Foreign.Ptr.Diff (peekDiffOff, pokeDiffOff)
+import GHC.Ix (Ix(..))
 
 import Geomancy.Elementwise (Elementwise(..))
 import Geomancy.Gl.Block (Block(..))
@@ -203,3 +204,34 @@ instance Block IVec4 where
   {-# INLINE write430 #-}
   {-# INLINE readPacked #-}
   {-# INLINE writePacked #-}
+
+instance Ix IVec4 where
+  {-# INLINE range #-}
+  range (l, u) =
+    withIVec4 l \l1 l2 l3 l4 ->
+      withIVec4 u \u1 u2 u3 u4 ->
+        ivec4
+          <$> range (l1, u1)
+          <*> range (l2, u2)
+          <*> range (l3, u3)
+          <*> range (l4, u4)
+
+  {-# INLINE unsafeIndex #-}
+  unsafeIndex (l, u) i =
+    withIVec4 l \l1 l2 l3 l4 ->
+      withIVec4 u \u1 u2 u3 u4 ->
+        withIVec4 i \i1 i2 i3 i4 ->
+          unsafeIndex (l4, u4) i4 + unsafeRangeSize (l4, u4) * (
+          unsafeIndex (l3, u3) i3 + unsafeRangeSize (l3, u3) * (
+          unsafeIndex (l2, u2) i2 + unsafeRangeSize (l2, u2) * (
+          unsafeIndex (l1, u1) i1)))
+
+  {-# INLINE inRange #-}
+  inRange (l, u) i =
+    withIVec4 l \l1 l2 l3 l4 ->
+      withIVec4 u \u1 u2 u3 u4 ->
+        withIVec4 i \i1 i2 i3 i4 ->
+          inRange (l1, u1) i1 &&
+          inRange (l2, u2) i2 &&
+          inRange (l3, u3) i3 &&
+          inRange (l4, u4) i4