+++ src/Geomancy/UVec2.hs
@@ -17,10 +17,11 @@ module Geomancy.UVec2
 
 import Control.DeepSeq (NFData(rnf))
 import Data.Coerce (Coercible, coerce)
-import Data.Word (Word32)
 import Data.MonoTraversable (Element, MonoFunctor(..), MonoPointed(..))
+import Data.Word (Word32)
 import Foreign (Storable(..))
 import Foreign.Ptr.Diff (peekDiffOff, pokeDiffOff)
+import GHC.Ix (Ix(..))
 
 import Geomancy.Elementwise (Elementwise(..))
 import Geomancy.Gl.Block (Block(..))
@@ -185,3 +186,28 @@ instance Block UVec2 where
   {-# INLINE write430 #-}
   {-# INLINE readPacked #-}
   {-# INLINE writePacked #-}
+
+instance Ix UVec2 where
+  {-# INLINE range #-}
+  range (l, u) =
+    withUVec2 l \l1 l2 ->
+      withUVec2 u \u1 u2 ->
+        uvec2
+          <$> range (l1, u1)
+          <*> range (l2, u2)
+
+  {-# INLINE unsafeIndex #-}
+  unsafeIndex (l, u) i =
+    withUVec2 l \l1 l2 ->
+      withUVec2 u \u1 u2 ->
+        withUVec2 i \i1 i2 ->
+          unsafeIndex (l2, u2) i2 + unsafeRangeSize (l2, u2) *
+          unsafeIndex (l1, u1) i1
+
+  {-# INLINE inRange #-}
+  inRange (l, u) i =
+    withUVec2 l \l1 l2 ->
+      withUVec2 u \u1 u2 ->
+        withUVec2 i \i1 i2 ->
+          inRange (l1, u1) i1 &&
+          inRange (l2, u2) i2