+++ src/Geomancy/IVec3.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BlockArguments #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE PatternSynonyms #-}
@@ -29,7 +30,7 @@ import Foreign.Ptr.Diff (peekDiffOff, pokeDiffOff)
 import GHC.Ix (Ix(..))
 
 import Geomancy.Elementwise (Elementwise(..))
-import Geomancy.Gl.Block (Block(..))
+import Graphics.Gl.Block (Block(..))
 
 data IVec3 = IVec3
   {-# UNPACK #-} !Int32
@@ -207,7 +208,7 @@ instance Storable Packed where
     <*> peekByteOff ptr 8
 
 instance Block IVec3 where
-  sizeOfPacked _  = 12
+  type PackedSize IVec3 = 12
   alignment140 _  = 16
   sizeOf140 _     = 16
   alignment430    = alignment140
@@ -219,7 +220,6 @@ instance Block IVec3 where
   write430    = write140
   readPacked  = read140
   writePacked = write140
-  {-# INLINE sizeOfPacked #-}
   {-# INLINE alignment140 #-}
   {-# INLINE sizeOf140 #-}
   {-# INLINE alignment430 #-}
@@ -233,7 +233,7 @@ instance Block IVec3 where
   {-# INLINE writePacked #-}
 
 instance Block Packed where
-  sizeOfPacked _  = 12
+  type PackedSize Packed = 12
   alignment140 _  = 16
   sizeOf140 _     = 16
   alignment430    = alignment140
@@ -245,7 +245,6 @@ instance Block Packed where
   write430    = write140
   readPacked  = read140
   writePacked = write140
-  {-# INLINE sizeOfPacked #-}
   {-# INLINE alignment140 #-}
   {-# INLINE sizeOf140 #-}
   {-# INLINE alignment430 #-}