+++ src/Geomancy/IVec4.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE BlockArguments #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -25,7 +26,7 @@ import Foreign.Ptr.Diff (peekDiffOff, pokeDiffOff)
 import GHC.Ix (Ix(..))
 
 import Geomancy.Elementwise (Elementwise(..))
-import Geomancy.Gl.Block (Block(..))
+import Graphics.Gl.Block (Block(..))
 
 data IVec4 = IVec4
   {-# UNPACK #-} !Int32
@@ -186,7 +187,7 @@ instance Storable IVec4 where
     <*> peekByteOff ptr 12
 
 instance Block IVec4 where
-  sizeOfPacked _  = 16
+  type PackedSize IVec4 = 16
   alignment140 _  = 16
   sizeOf140 _     = 16
   alignment430    = alignment140
@@ -198,7 +199,6 @@ instance Block IVec4 where
   write430    = write140
   readPacked  = read140
   writePacked = write140
-  {-# INLINE sizeOfPacked #-}
   {-# INLINE alignment140 #-}
   {-# INLINE sizeOf140 #-}
   {-# INLINE alignment430 #-}