+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Added `convert` function to vector modules to facilitate type-changing operations like rounding.
 * Added `Ix` instances for integral vectors.
+* Added `dot` for integral vectors.
 
 ## 0.2.5.0